refactor: Split reducers into individual files
